// apps/nestjs-backend/src/config/config-key.enum.ts
export enum ConfigKey {
  NODE_ENV = 'NODE_ENV',
  FRONTEND_HOST = 'FRONTEND_HOST',
  PORT = 'PORT',
  ENABLE_SWAGGER = 'ENABLE_SWAGGER',
  ENABLE_DEMO_SEED = 'ENABLE_DEMO_SEED',

  POSTGRES_TIMEZONE = 'POSTGRES_TIMEZONE',
  POSTGRES_DB_NAME = 'POSTGRES_DB_NAME',
  POSTGRES_PASSWORD = 'POSTGRES_PASSWORD',
  POSTGRES_PORT = 'POSTGRES_PORT',
  POSTGRES_HOST = 'POSTGRES_HOST',
  POSTGRES_USER = 'POSTGRES_USER',
  POSTGRES_DEBUG_MODE = 'POSTGRES_DEBUG_MODE',

  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_PASSWORD = 'REDIS_PASSWORD',
}